@@ -1,4 +1,4 @@
-class Requests < ActiveRecord::Migration
+class Requests < ActiveRecord::Migration[6.1]
   def change
     create_table :requests do |t|
       t.string :method, null: false
